foundry: avoid build installation if no changes occurred
This is an optimization to avoid re-installing the project if we have not
detected any files to have changed within the project. You can always
force a build using the build button, so this seems a better default for
the majority case which is that all files are within the project directory.