code: add API for buffer addins to settle
This gives the buffer addins a way to say they need more time to finish
an in-flight operation before the buffer is marked as "settled". Doing
so allows things like "save" to settle before proceeding with reading the
buffer contents.

This only implements the addin scaffolding. Additional work in IdeBuffer
is needed to consume this.