project-tree: add project-tree to editor workspaces
This allows use of the project-tree from editor workspace windows. However,
it does not include project-related items unless a project is loaded in the
IdeContext.

The workdir is used, which is the top-level directory found when opening
files from the command-line (or the project workdir in the case a project
was opened).