pipeline: add arch helper
We can't always trust the arch specifically in a runtime object (as that
might be one matching the host, not the target device). So add a helper
that will check the device first and fallback to runtime.

In the future, we should probably rename the flatpak runtimes to only
use the name/version instead of arch too.