safety: avoid cast of GDestroyNotify
This allows us to introspect the type of the return pointer if
the callback is suitable to handle that parameter type.

In some cases, we need to use (ide_task_return_pointer)() which
could be worked around if we really cared to, but I'd rather not
for now until we deem it more useful to return NULL free
functions (or data).