code-index: avoid indexing on un-prepared pipelines
If the current build pipeline has not configured, or we don't
have a project, refuse to begin indexing the project.

This fixes a warning about builddir from the pipeline when it
races to be setup. Also, it reduced the chances we do indexing
immediately at startup for un-prepared build systems.