flatpak: query user before downloading runtimes
This uses the new installation dialog as part of the async
workflow to initialize the build environment.

The user is queried to confirm the installation of sdk/runtime
as requested by the manifest for the target architecture.

It would be nice if we could add some more info to the
installation dialog in the future (like size, if it's missing
on the peer, etc)