sourceview: make completion window translucent on ctrl toggle
This requires libdazzle 3.33.0 for shortcut fixes.

When pressing Control while the completion window is present, it will
become translucent. Pressing Control a second time will restore that
visibility.

Whenever the completion window is displayed, it will be forced back to
full opacity.

This allows for seeing text around the completion window without having
to dismiss the completion window fully.