git: port git plugin API to use gnome-builder-git daemon
This ports the various plugin API for Builder's git integration to use
the new gnome-builder-git daemon. The daemon is spawned and uses the
DBus marshalling format (without a DBus daemon) over STDIN/STDOUT pipes
to the subprocess.

The daemon itself manages most of the state now and allows for some
more threading from DBus proxies in the Builder code-base.