gui: add IdeSearchButton widget
This widget looks and acts like a GtkButton, but switches to a GtkEntry
(using DzlSuggestionEntry) with a transition, keeping the magnifying
glass icon in the same position. Only theme styling for Adwaita ensures
this, other themes would need tweaking for likely pixel alignment.

Using this widget for search will allow saving some space in the header
bar when searches are not being performed.