gui: switch workspaces to IdeSearchButton
This uses the new IdeSearchButton widget to allow the headerbar to
shrink more, and then expand as necessary for search.

One possibly interesting future direction for this widget is for it
to expand using a new window layered over the entry so that no space
is required from a headerbar.