clang: locate C++ standard includes using clang++
This attempts to discover the location of C++ includes when we detect
a C++ file is used (by sniffing -std=c++ or similar). We may need to add
-xc++ to, but so far with what I've tested, -std=*++* is enough.

This is neccessary because build systems don't give us enough information
about how to compile the C++ project and clang++ (and g++) support
compiling with non-default C++ standard library. (Such as you might use
when cross-compiling targets).

I imagine that we will need to unwrap/revert some of this when we end up
supporting more cross-architecture development, but that remains murky
for now and this gets things working.

Fixes #900