vagrant: prototype a vagrant runtime
This still needs work on various fronts, but it's good enough to
iterate upon now. To make some things we need fast we will need
to capture some information at load time.

For example, we expect "contains_program_in_path" to be fast, so
we will want to ls /usr/bin, /bin, etc and cache the result.

We also wont have the ability to pass FDs across the process
boundary, so we'll have to introduce new API in
IdeSubprocessLauncher for something like "can_pass_fd". That way
subsystems like GDB can use alternate strategies to spawn a
debugger.

Furthermore, we need to extend the build pipeline to sync code
to the domain. We can look at adding an IDE_BUILD_PHASE_PUSH
that happens before IDE_BUILD_PHASE_CONFIGURE. A vagrant plugin
would likely want to rsync code here.

There will be fallout for a number of other things too, such as
how do we do exports, execution with display, etc.

We might also want to add a Containers project-tree node so that
some basic management of containers can be done.