meson: make testprovider load tests when build.ninja changes
This avoids forcing the pipeline to build to load meson tests. To do so,
we need to watch build.ninja for changes so that we can run "meson
introspect" as needed.

We also need to always load them after the first pipeline request (which
is generally to CONFIGURE phase).

The side-benefit to this is that we can avoid advancing to the CONFIGURE
phase of the build pipeline on startup.