notification: avoid use of buildmanager from pipeline addin
This notification addin is a pipeline addin, and therefore should try to
focus solely on the pipeline itself. Otherwise we risk multiple objects
in flight changing state while the second pipeline is in cleanup.

This also tries to improve the cleanup of notifications to the shell when
the pipeline is destroyed.

Furthermore, we attempt to show more notification during the initial build
so that the user has some idea what is going on.

When paired with the meson-introspect fix in the previous commit, we should
be a lot less chatty during startup, yet fully informative about what is
going on.

Fixes #911