flatpak: simplify discovery of programs in path
We do not need to run `which` since we have the deploy-dir and all runtimes
currently put things in /usr/bin. We can just iterate known PATH entries
(currently /app/bin:/usr/bin, but /app/bin does not matter) instead of
relying on process execution which might round-trip to the host anyway.