io: specify O_NOCTTY during child PTY creation
When creating the child side of a PTY pair, we need to ensure that we do
not allow the PTY to become the controlling terminal for the current
process. We won't have one of those when run from shell, and we very much
don't want this to become ours (or else we'll SIGHUP when it is closed).

Fixes #971