vcs: initialize the struct directly
Instead of initializing a struct to its zero value and assign values to
its member afterwards, we can set values for members directly in the
initializer. This avoids the missing braces warning for clang when the
first member is a struct or an union whose first member is a pointer.