c-pack: improve indentation with tabs
This improves how we indent when we have different values for indent-width
and tab-width. This is useful with GNU style indentation when we are
allowed to use a real \t instead of spaces.

For this to work correctly, it also requires the following from
GNOME/gtksourceview#64.

Currently Builder does not expose separate GSettings for tab-width vs
indent-width, but that may change in the future. You can still set these
independently with .editorconfig though.

Fixes #309