create-project: set Rust as a preferred language
The first bar of 4 languages is this maintainers preferred languages for
new projects. Given that so much new code is happening in Rust, it seems
prudent for us to switch now so that we can ease the burden on starting
projects in a relatively safer language.

Just because your language of choice isn't in the simplified row, doesn't
mean you can't keep using it. But please don't take this as any sort of
official message from GNOME, just my choice as an oppinionated
author and maintainer.

Use what you like, and build proper security boundaries into your
application's architecture no matter what language you choose.