core: remap various architectures
We want to keep these consistent so that they map to external tooling no
matter what the host system claims. We might need to do this in a couple
other places in the future, but for now, given the design of Builder, this
is probably the widest net to cast.

Fixes #1061
Fixes #1070