snippets: remove else from C snippets
This one is really annoying for cases where you don't want brackets. And
we already get things right when you do want brackets because you can
just hit {, have the secondary } auto-added, and hit enter again to be
placed in the right spot.

So all in all, I think having this snippet was worse than not.