terminal: follow filename hyperlink on button press
It doesn't really make sense to require CONTRO_MASK for following links
to files here. First off, people are already used to underlines for
hyperlinks in browsers, we show a hand, and control click isn't really
a pattern for browser links.

Since we're already going through all that work, just follow the link.
We do risk this colliding with interactive terminal sessions like curses
applications, but given our scope, I'm okay with that.

Reported-by: Lennart Poettering
Signed-off-by: Christian Hergert <chergert@redhat.com>