lsp: allow clients to reply to workspace/configuration
If a workspace/configuration call is made by the language
server to our IdeLspClient, we want to reply with something
useful to the peer.

Since this content can be different for every LSP, we'll defer
to the plugin to provide the content. Some params we'll want to
make easier for the plugin to determine, but this seems like the
quickest first step to make that possible. We can add additional
abstractions for the others later.

Related #1074