lsp: implement breaking change from LSP 3
LSP 3 no longer has a containing "location" node. So instead
we should just check for the Range in the immediate object.

It might be nice to handle things going backwards, but given
how quick the other editors are moving forward, it's probably
more likely that LSPs move forward quickly too.