lsp: add support for workspace/applyEdit
This implements the simple form "changes" for workspace/applyEdit. It
would be nice to add support for documentChanges at some point, but I
think we want something to help with that first to deal with all the
file changes that can occur.

This at least implements the minimum required to say we support it.

Related #1088