html-preview: update markdown CSS from upstream
This improves things to actually be readable in night mode, but does
unfortunately result in a white background. This is still better than
not being able to read anything.

If we want something that tracks media queries propertly, that would be
nice but is likely a fair amount of work for someone.

Fixes #1093