i18n: remove context from shortcut menus
These cannot actually be translated because it requires the use of
g_dpgettext2() to provide the context for the translation. These should
have never been changed to NC_() to begin with. If the context is truly
necessary, translators will need to find an alternate way to provide this
by not using the same translation for every msgid using the same wording.