copyright: add copyright plugin that can update copyrights
This plugin, when enabled, can update copyright lines in the headers of
your project. A toggle is provided in Preferences and defaults to false.

This works by scanning for your real name, provided by g_get_real_name(),
and then looking for years using a 4 digit regex. If a range is found,
the range will be updated. Otherwise the first number will be updated.

There are certainly things this will break on, but the limited set of
things I've tested on work well. If you have something it breaks on, let
me know and I'm sure we can accommodate.