symbol-tree: avoid jumping between files when movin to symbol
We do not want to rely on the IdeEditorSurface to jump to a location here
because it means that we might not activate the same IdePage that we are
currently viewing.

Instead, this tries to jump usin the local frame, and then falls back to
the editor surface (just in case we raced to complete the operation with
a page switch).

Related: #723