editor: maintain line/column after reloading file
We want to try to keep the cursor in as close of position as we can to the
previous form when reloading the file. This tries to maintain that but
will fallback to the latest line (assuming line is past the last line)
and then the column up to the last column (assuming line_offset is past
the end of the line).

Fixes #1211