plugins/podman: Streamline runtime launcher creation
'gbp_flatpak_runtime_create_launcher' has a runtime id check at the
beginning. This adds it to 'gbp_podman_runtime_create_launcher' to unify
the implementation.

The 'run-on-host' == 'TRUE' is redundant in the case of Podman's runtime
launcher because it is always executed on the host (implementation
detail of 'gbp_podman_subprocess_launcher_spawn').