plugins/podman: Fix envvar setup in launcher
The problem with setting envvars could be encountered when 'clear_env'
was set to 'False' in the runtime launcher. In some cases, the envvar
did not exist but the index still got incremented, causing problems in
the 'ide_subprocess_launcher_push_argv' function. With this the index
is only incremented when the said envvar is set.