code: chain result for competing get_symbol_at_location
If we get two requests to get the symbol at a given location, at the same
time, without any changes to the buffer, then we want to reuse the result
for both requests.

This implements that to cut down on the overhead to show the hover window
to users (which required usin this from at least two plugins).

Also note we only show things after all hover providers have returned, so
some work could be done to update those as results come in.

Fixes #1313