editor: add missing ids
This is rather annoying, but for how menu merging works we need the id
within the section item as well as in the <section id="..">. Some work
may be able to be done in dazzle to improve this longer term. However,
I think the real goal should be to have menu merging in GTK 4 directly.