project-tree: add gbp_project_tree_reveal()
This helper handles the asynchronous nature of expanding the file tree to
find a given file or directory within the project. The tree will be
expanded up to that node and the node will be selected and scrolled into
view of the user.

We used ot have something like this before movin to IdeTree, and this
restores that functionality.

Related #1320