run: Stop showing full GLib debug messages when not debugging
Currently when running an application, Builder enables all GLib debug
messages, which means that it always spams the logs, which makes it hard
to see actual printing inserted in work-in-progress code.

So this commit stops enabling debug messages with the Run button, but
still shows them when running with the debugger, so there's still an
easy way to see those.

Additionally, it's still possible to use the Runtime Environment section
in the Build Preferences to enable those again.

Fixes #1270