project-file: Show overridden symbolic icon for the MIME type
Currently the icon associated for a MIME type was meant to either be
picked either from the custom overridden icons included in Builder,
mostly programming languages related, or from the default theme ones, if
available.

But there was a bug where the content type used to choose the custom
icon bundled in Builder only checked for the MIME type by file name
(like if the file ends with .xml), by using the "fast" content type
detection. This made pretty much every custom icons not to be used as
not correctly detected it, and hence very few icons were shown in the
project tree sidebar.

So this commit fixes this by using the normal content type detection
instead of the "fast".