gui: hide three-finger-swipe gesture behind shift check
In GNOME 40 Shell will have three-finger-swipe gestures that collide with
applications. This switches this to a shift+three finger swipe gesture
for now to move frames.

Given that Builder is the only applications using this we are probably
better off to let the Shell own three-finger gestures.

We can still improve on two-finger swipes once we go to GTK 4 for moving
between frame stacks quickly.

Fixes #1371