ls: Don't close it after file opening when opened from the project tree
Currently when choosing a file in a folder browser opened from the
project tree menu, the folder browser is closed.

But this behavior is inconvenient when someone wants to use the folder
browser as a persistent and pane-able view of a folder.

So this commit changes this behavior to let the folder browser opened
after opening a file. It is still possible to use it as a vim-like
directory listing, by using the "Switch to Folder" item in the right
click menu in an editor view.