buildui-omnibar: Allow buttons homogeonous property to be translated
This property was meant to be translatable, as the comment on the line
before indicates. It'll allow translators to manually toggle whether
the buttons in the omnibar should he homogeonous, because in certain
languages it takes a really big width which means that the different
labels in the popover and buttons are spread in long distances.

So this commit makes it translatable.