project-tree: Clarify "Reveal in Project" label
Currently it is unclear where this "Project" leaves in the UI, and hence
what this menu item will do. In fact, this item selects the currently
shown file in the project tree, located on the left sidebar.

So this commit clarifies this label text, by simply specifying it's the
"Project Tree", which is clearer.