icons-mimetype: Fix rendering of the changelog icon
Currently the rendered text-x-changelog-symbolic.svg PNGs have some sort
of line of circles, while there's only three separated circles in the
original SVG.

This is because there's some circles in the SVG that are using fill=none
to make them invisible. It must have been while iterating on the icon,
maybe testing with four dots+rectangles.

So this commit fixes the original SVG, by simply removing the invisible
circles, and regenerates the corresponding PNGs.