Stop using deprecated gtk_style_context_get_*color() functions
Currently in various place, gtk_style_context_get_color() and
gtk_style_context_get_background_color() are used to get and replicate
the color a color to an other place.

The later is deprecated in GTK3 and removed in GTK4. So this commit
uses other means to get the background color. Since the get_color()
function is quite inexpressive in the sense that you don't really know
which color you'll get (like the get_background_color()), this commit
also replaces those uses by more saner ones.