ls: Activate rows on single click
Currently all rows in an ls page needs to be double-clicked to open
the file or the directory.

This is inconsistent with both the project tree and opened files list
in the editor's sidebar, which both activate rows on single click
because it's way more convenient, and because there's no point in using
double click when single click does it better.

So now rows gets activated on single click rather than double click.