editor: Clarify Frame's menu section Close item text
Currently the Close item in a frame header's menu is under a "Frame"
menu section.

But it isn't clear whether this item will close just the page, keeping
the frame opened if there's still some pages in it, or whether it'll
close all the pages in the frame and the frame itself.

Since this button is intended to close the whole frame, then clarify the
item's text accordingly.