intelephense: Add intelephense lsp client implementation
This introduces an lsp client implementation for the intelephense
(https://intelephense.com/) language server for PHP.

The plugin supports most of the functionality in the language server,
but not everything is accessible without a licence key for intelephense.
There is currently no way to provide this license key except modifying the
plugin itself.