session: Make IdeSession handle plumbing so addins only handle content
Currently the IdeSession provides a very bare session handling, where
pretty much everything needs to be done by the addins, which represents
a non-trivial amount of work, meaning that it's hard to have several
addins.

This also introduces problems like not being able to have several addins
restoring to the workspace's grid, because the restore order isn't
guaranteed so pages will end up in the wrong places (because the
restoring isn't handled at one place but at several ones).

So this commit makes the IdeSession handle all this plumbing in the
scene, so addins only have to handle the content of the IdePage (e.g.
set editor page to URI foo://path/to/bar, with search set to "foobaz"),
which makes them much more approachable to write. This also solves the
issue of restore order, since the IdeSession can first ask all addins
to restore the pages, then sort it by their position in the grid, then
add them back in their right position in the grid.