beautifier: Print formatter error on its own line
Currently when a formatter outputs an error message giving the location
of the error, like so:
foo bar baz
    ^^^
    \---------- formatter error message
It doesn't fit nicely because the first line is after the foreword that
is printed just before on the same line.

So this commit shows the formatter error message starting on its own
line, so this kind of error message is properly displayed.