gstyle: Drop X11-only eyedropper in favor of portal-based color picker
Currently the color picker (eyedropper) works correctly only on X11,
since there's no way integrated into Wayland to take screenshots/pick a
color.

So this commit switches it to use the portal through libportal, which
means that the system is responsible for providing a convenient
eyedropper/color picker (which includes showing a bigger zone of the
color at the cursor, a "zoom" of it) while being able to use it with
both X11 and Wayland.