refactor: rename can_install to provides
The reasoning behind this is as follows: A runtime might
need more time to finish setting up than is allowed in the
synch `_load`. In that case the provider extension will want
to indicate that it is able to provide the runtime for
the configured runtime_id, but that it's not yet ready.

The manager can then start the bootstrap process and allow
the provider to finish setting up, and then let the manager
know that it is done. So that the build manager knows when
the pipeline is configured and ready to go.