editor: Add shortcut to sort selected lines
Currently if someone needs to sort lines like for e.g. a list of files,
there's no shortcut to do it so one needs to go find the correct menu
item with the right click menu.

Since there's already a lot of other shortcuts for those kind of things,
it makes sense to add a shortcut for it.