update-manager: Make sure there's only one update notification
Currently when running Builder nightly there's frequently updates, so
during a full day it might happen that the flatpak portal sends an
"update available" signal several times.

But the plugin currently doesn't replace the old update notification
with a newer one, so they start to pile up in the build popover.

So this commit stops early in the "update available" signal callback in
case a notification is already there.