threading: Make ide_subprocess_launcher_set_argv take const arg list
Everywhere in both GSubprocessLauncher and IdeSubprocessLauncher are
arguments passed as const char * or const char * const *, to make it
easier to pass arguments directly as string literals, like "foo", or
(const char * const *){"foo", "bar", "baz", NULL}.

For this reason, it's a good idea to make set_argv() take fully const
argv. This doesn't break API/ABI in any way, but it'll throw a warning
for places where it was explicitely casted to (gchar **), so this commit
also fixes the only occurence in Builder's code base where a warning was
thrown.