flatpak: Add html-tidy module
Currently when using the HTML/XML beautifier with the flatpak, an error
appears saying "tidy" isn't found.

This is because it's not built in the manifest, like is e.g. uncrustify,
so it can't be found.

So this commit builds the git version of html-tidy in the flatpak
manifest, because the last release is more than three years old.