core: add _ide_host_environ()
This private API attempts to extract the environment from the host so that
we can proxy necessary variables into new containers. In particular, this
is useful to get the correct XAUTHORITY, DISPLAY, and other things which
might be different in the Flatpak runtime.

Generally, you wouldn't do this, but since we are a developer tool that is
mostly just using Flatpak for packaging, it helps with the integration
tooling for podman, toolbox, etc.