terminal: use environment from host for terminals and runners
We want to be closer to the host system environment than what we see
ourselves in the Flatpak runtime. This allows us to copy some variables
from the host as the user would see them natively.

This allows me to now run things like gtk4-demo inside a podman that is
used as a SDK runtime.