gui: Bind frame popover open pages icon by gicon instead of icon-name
Currently editor pages have their page icon set to the one found for
the corresponding MIME type of the file.

While those icons are correctly shown under "Opened Pages" in the left
sidebar, they are only shown with the default text/plain MIME icon in the
frame header's Open Pages popover, so per-MIME type icon support is left
unfinished.

So this commit fixes it by changing from a binding of the icon-name
to a binding of the (g)icon, because the icon-name property is only a
shorthand to set the gicon, but icons at the end of the day are always
set in the (g)icon property even when setting icon-name. This is
especially important since ide_g_content_type_get_symbolic_icon gives a
GIcon, not an icon name. That's why editor pages were shown with the
text/plain MIME icon in the popover, as the icon was first set with
icon-name before setting it with the gicon associated with the file's
MIME type.