flatpak: handle incorrect ref in SDK extensions
This fixes an issue where an extension might say that it extends a runtime
for an architecture other than the architecture it advertises as its
runtime.

Notably, org.freedesktop.Sdk.Compat.i386/x86_64/19.08 says that it is an
extension of ref=runtime/org.freedesktop.Sdk/i386/19.08 which can't be the
case (since it's x86_64).

Fixes #1437