flatpak: allow cancellation of transfer
Quite frankly, this can cause an issue where we get the repository into a
state that needs repair, but I think it's better that we at least do what
the user said than nothing at all. Data can be expensive on some networks.

Ideally this should be addressed in libflatpak at:

  https://github.com/flatpak/flatpak/issues/4258