project-tree: Reveal left dock when using "Reveal in Project Tree"
Currently if the left dock (where the project tree lives) is hidden,
then using the Reveal in Project Tree feature won't make it reveal from
the side, which is annoying because if you want to reveal a file in the
project tree then you'll surely want to actually see it, which isn't the
case here as the dock isn't revealed. It means you need to press F9
to manually reveal it instead of having it done automatically.

So this commit removes this annoyance by revealing the left dock when
revealing a file in the project tree.