buildui: Enhance translators comment with CAPS
It seems like quite a few language teams have translated wrongly the 'true'
homogeonous value, although there clearly is a warning saying it shouldn't be
localized to the target language. Just run grep '^msgid "true"' -A 1 *
--no-filename --no-group-separator | sort | uniq | grep -vE '(false|true)' in
the po/ directory will show you that 11 language currently have a broken button
layout in the popover (well, not really broken but it won't have an
appropriatate sizing).

I doubt it will change much to it, but I changed the true to True so it
invalidates the translation and hopefully they'll see the big caps translator
comment. :)