open-with-external: Split from project-tree plugin to separate plugin
Currently it has a specific GAction which is only usable from the
Open With => External Program context menu entry in the project tree.

But by making it a separate plugin that handles opening through a
workbench addin, it can have a specific very low priority to act as a
fallback anywhere a file opening operation takes place: from the Ctrl+.
search bar and from the project tree. This means in particular that
files that can't be opened properly by Builder currently, like sounds,
images and archives will instead be opened with an external program
instead of the directory browser view (which is a bug anyway).