gui: Cancel previous search when typing new text in search button
Currently when having quite a few disk/CPU activity from another
process (like compiling or just running `yes`), the global file
search has a fair bit of latency when typing text, as the previous
searchs are having a hard time with IO. So they tend to "pile up"
when typing new characters, which means that there's a few seconds
where the search results will change several times as search results
start arriving.

To prevent this sluggish search, this commit now cancels the
previous search request in favor of a fresher one with the
newly typed text.

Also fixes a FIXME, where search errors are now reported in the
messages pane instead of hidden in the logs.