buildui: Add mnemonics to buttons and make small buttons focusable
For no good reasons the small buttons with an icon had can_focus set to
False, which makes them impossible to focus with the keyboard. There's
a lot of useless can_focus = False on labels/images in this file (and
probably others), but unless they break the keynav there's no point in
making huge diffs for that as it'll move all around when porting to
GTK 4.