vcsui: Drop dead code for project tree VCS indicator color
vcs-added and vcs-changed (initially found in the CSS) were only
mentioned in the vcsui tree addin, where the only state that changed,
the two rgba colors in the struct, weren't ever used.

Effectively those were dead code, so this commit removes them. It's
likely that the colored status indicator feature was dropped at one
point, but that this part of the code was missed.