themes: Explicitly specify i-wanna-be-listbox regular background color
While having a transparent background works fine for the most simple
tree views, when there's drag and drop involved it breaks because the
dragged row has a black background, so the text is unreadable and it
is expected that the row keeps its "real" background color there.

So to fix this, this commit sets the background color to the
theme_base_color.