tree: Make sure that the focused row is changed too when selecting one
Currently if the selected row and keyboard input focus is on row A, and
one tries to right click row B to show the context menu, then it'll try
to select the row with ide_tree_select(). But the keyboard focus isn't
actually changed, only the tree model selected row, so pressing the down
arrow won't select row C but it'll select row B instead, as the focus
was still on row A although the selected row was B.

So this commit makes the keyboard focus change at the same time as a row
is selected, in ide_tree_select().