manifest: Set devhelp's profile to stable
Two months ago devhelp gained support for separate devel and stable installations,
so the installed app icon depends on it. Since we're using devhelp's symbolic icon
as IdePage icon, it wasn't able to get the icon because the default profile was
devel, so the installed icon was org.gnome.Devhelp.Devel-symbolic.svg, instead
of the non-devel variant.

So this commit switches to stable variant to fix the icon.