gui+editor: Make row close button keep its focus outline
Currently when navigating in the frame open pages popover or the open pages
in the sidebar, moving the focus on one of the close buttons can be confusing
because the focus outline was disabled, so it appears like the focus was lost
somewhere, even though pressing Enter/Space will close the row.

So to fix this, the CSS that disabled the outline was removed, and the buttons
now have valign=center so that the focus outline stays square with a bit of
padding.