gui: Use a single run button instead of GtkStack -like double hidden buttons
Currently when clicking the run button, a slight border radius animation due
to the linked style class happens. This is because there's actually two buttons,
one for the "stop build" and one to "run build".

While we could just be disabling transitions to avoid this issue, there's not
much reasons to have two buttons like that. We could be using a stack instead,
but that won't work with the linked style class, which basically expects
the "box.linked > button" CSS selector here. So instead, this commit listens
for changes of "busy state" of the run manager, and the run handler change
to change both the Run button icon and the action name triggered by the button.