gui: Make IdeSession pre API rework migration use a definite GVariantBuilder type
Currently if opening a project which used the old session.state format (pre API
rework) that doesn't contain any page (just an empty workspace basically), the code
would spew
sys:1: Warning: g_variant_builder_end: assertion '!GVSB(builder)->uniform_item_types
|| GVSB(builder)->prev_item_type != NULL || g_variant_type_is_definite
(GVSB(builder)->type)' failed
with a few other related warnings that followed.

This is due to the current code that initialized the array of pages' states with
an indefinite (i.e. not known fully) array type. Usually the GVariantBuilder
infers the type of the elements in the array by looking at the type of the first
added item, with g_variant_builder_add(_value) function. But since no item
is added, the GVariantBuilder doesn't know what's the definite type of the array
when g_variant_builder_end() is called, it returns NULL and emits a warning,
because there's no way it can build the (empty) array in that case.

So this commit initializes the GVariantBuilder to the full definite type aa{sv},
since we know what's supposed to be in the GVariant at the end.