gui: Save state to disk even when there weren't any page's state to save
Currently (after session API rework) if a workspace doesn't have any page opened,
nothing's saved at all in the session.gvariant file. This is because it is only
done when the asynchronous operations count reaches zero _and_ only from the
per-page saving async callbacks. So since there isn't any page to save, the
callback isn't called and the state is never saved to disk, even if it exists as
as empty GVariant array.

So to fix this, this commit splits the state saving function into a separate one
that strictly handles only saving to disk, and call that save to disk function
when there's no pages to save.