vls: bind LSP client lazily for search
This ensures that the client is not spawned unless another component has
caused it to.

That does have the side-effect of requiring the user to actually open
something that requires vala before search results are displayed. Not the
end of the world, but in the future it might be nicer to sniff this out.

Fixes #1476